requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
Pillow
drf-yasg
python-decouple
django-filter
django-allauth
gunicorn
whitenoise
dj-database-url
django-cors-headers
celery 
redis 
drf_social_oauth2
google-auth

